demo/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0
python-dotenv==1.0.0
marshmallow==3.20.1
